Updates on the code
Added a wiener filter that can be used for the y and z axis, since these experience the most disturbance from the floor. This is only to show what the signal should truly look like, but do not use this for creating a model, since the disturbance from the floor actually makes the measurement (mainly in the y axis) more WSS. Next, also a code to fit an ARMA and, although it does not really work, an ARMAX model. Perhaps this can later be used. Also, somewhere in the beginning, a code was added to desample a measurement from for example 2000 Hz to 1000 Hz, which might make it easier to fit a model.