Updated the code with cross correlation
I have added a code that shows the cross correlation between two axes (so x and y for the same measurement for example). It is very clearly visible if a measurement is OK or NOK, since a NOK has very high peaks (mainly above 0.2 at lag of 0) and also has peaks at lags that are not equal to zero, mainly above 0.05. Also NOK measurements are symmetric, which the OK not always has, so that is also an indicator for a NOK measurement