Add files via upload
Fit an AR model on M01/OP7. Result residuals are white noise. applying the filter on the bad data of MO1 gives colored risiduals, also squared error is way bigger. Crosscorrelation between M01 and M02, suggest not correlation values at all lags are very small, same for M01 and M03. However i used a command, not sure if this gives the correct result i am looking into another way to compute and i will compare. Also when i do the residuals test on the good and bad of the other machines with the model fitted on M01 the residuals are also colored.